+++ index.js
@@ -16,128 +16,127 @@ function hashtag_text(tokens, idx) {
   return '#' + tokens[idx].content;
 }
 
+//////////////////////////////////////////////////////////////////////////
 
-function findHashtag(string, start, end, validChar, specialTags) {
-  var tagName, pos;
-  // we need at least one char for a tag
-  if (start + 1 >= end) {
-    return null;
-  }
-
-  // check if we have special tags
-  if (typeof specialTags !== 'undefined') {
-    var match = specialTags.exec(string.slice(start + 1, end));
-    // we found a special tag
-    if (match !== null && match.index === 0) {
-      tagName = match[0];
-      pos = start + 1 + tagName.length;
-      return { tag: tagName, pos: pos };
-    }
-  }
-
-  if (!validChar.test(string.charAt(start + 1))) {
-    // no special tag and no valid char -> no hashtag
-    return null;
-  }
-
-  // there is a regular tag
-  pos = start + 1;
-  while (pos < end && validChar.test(string.charAt(pos))) {
-    pos++;
-  }
-  tagName = string.slice(start + 1, pos);
-  return { tag: tagName, pos: pos };
-}
-
+function isLinkOpen(str)  { return /^<a[>\s]/i.test(str); }
+function isLinkClose(str) { return /^<\/a\s*>/i.test(str); }
 
 module.exports = function hashtag_plugin(md, options) {
-  function hashtag(state, silent) {
-    var tagName,
-        max = state.posMax,
-        start = state.pos,
+
+  var arrayReplaceAt = md.utils.arrayReplaceAt;
+
+  function hashtag(state) {
+    var i, j, l, m,
+        tagName,
+        token,
+        tokens,
+        blockTokens = state.tokens,
+        htmlLinkLevel,
+        matches,
+        text,
+        nodes,
         pos,
-        preceedingChar = /\s/,
-        hashtagChar    = /\w/,
-        specialTags;
+        level,
+        regex,
+        preceding     = '^|\\s',
+        hashtagRegExp = '\\w+';
 
     if (options) {
-      if (typeof options.preceedingChar !== 'undefined') {
-        preceedingChar = options.preceedingChar;
+      if (typeof options.preceding !== 'undefined') {
+        preceding = options.preceding;
       }
-      if (typeof options.hashtagChar !== 'undefined') {
-        hashtagChar = options.hashtagChar;
-      }
-      if (typeof options.specialTags !== 'undefined') {
-        specialTags = options.specialTags;
+      if (typeof options.hashtagRegExp !== 'undefined') {
+        hashtagRegExp = options.hashtagRegExp;
       }
     }
 
-    if (state.src.charCodeAt(start) !== 0x23/* # */) { return false; }
-    // either the tag is at the beginning of the line or
-    // there has to be a preceeding char in front of the tag
-    if (start > 0 && !preceedingChar.test(state.src.charAt(start - 1))) {
-      return false;
+    regex = new RegExp('(' + preceding + ')#(' + hashtagRegExp + ')', 'g');
+
+    for (j = 0, l = blockTokens.length; j < l; j++) {
+      if (blockTokens[j].type !== 'inline') { continue; }
+      tokens = blockTokens[j].children;
+      htmlLinkLevel = 0;
+
+      for (i = tokens.length - 1; i >= 0; i--) {
+        token = tokens[i];
+
+        // skip content of markdown links
+        if (token.type === 'link_close') {
+          i--;
+          while (tokens[i].level !== token.level && tokens[i].type !== 'link_open') {
+            i--;
+          }
+          continue;
+        }
+
+        // skip content of html links
+        if (token.type === 'html_inline') {
+          // we are going backwards, so isLinkOpen shows end of link
+          if (isLinkOpen(token.content) && htmlLinkLevel > 0) {
+            htmlLinkLevel--;
+          }
+          if (isLinkClose(token.content)) {
+            htmlLinkLevel++;
+          }
+        }
+        if (htmlLinkLevel > 0) { continue; }
+
+        if (token.type !== 'text') { continue; }
+
+        // find hashtags
+        text = token.content;
+        matches = text.match(regex);
+        if (matches === null) { continue; }
+        nodes = [];
+        level = token.level;
+
+        for (m = 0; m < matches.length; m++) {
+          tagName = matches[m].split('#', 2)[1];
+
+          pos = text.indexOf(tagName);
+
+          if (pos > 0) {
+            nodes.push({
+              type: 'text',
+              // char at pos-1 is '#'
+              content: text.slice(0, pos - 1),
+              level: level
+            });
+          }
+          nodes.push({
+            type: 'hashtag_open',
+            content: tagName,
+            level: level++
+          });
+          nodes.push({
+            type: 'hashtag_text',
+            content: tagName,
+            level: level
+          });
+          nodes.push({
+            type: 'hashtag_close',
+            level: --level
+          });
+          text = text.slice(pos + tagName.length);
+        }
+
+        if (text.length > 0) {
+          nodes.push({
+            type: 'text',
+            content: text,
+            level: state.level
+          });
+        }
+
+        // replace current node
+        tokens = arrayReplaceAt(tokens, i, nodes);
+        blockTokens[j].children = tokens;
+      }
     }
-
-    var result = findHashtag(state.src, start, max, hashtagChar, specialTags);
-
-    if (result === null) { return false; }
-
-    pos = result.pos;
-    tagName = result.tag;
-
-    if (!silent) {
-      state.pos = start + 1;
-      state.posMax = max;
-      state.push({
-        type: 'hashtag_open',
-        level: state.level++,
-        content: tagName
-      });
-      state.push({
-        type: 'hashtag_text',
-        level: state.level,
-        content: tagName
-      });
-      state.push({ type: 'hashtag_close', level: --state.level });
-    }
-
-    state.pos = pos;
-    state.posMax = max;
-    return true;
   }
 
-  md.inline.ruler.after('emphasis', 'hashtag', hashtag);
+  md.core.ruler.after('inline', 'hashtag', hashtag);
   md.renderer.rules.hashtag_open  = hashtag_open;
   md.renderer.rules.hashtag_text  = hashtag_text;
   md.renderer.rules.hashtag_close = hashtag_close;
-
-  // TODO check if there is a cleaner solution
-  // make sure there are no hashtag links inside of links
-  md.core.ruler.push('remove_hashtag_inside_links', function(state) {
-    var i, blkIdx, inlineTokens, openLink;
-    for (blkIdx = 0; blkIdx < state.tokens.length; blkIdx++) {
-      if (state.tokens[blkIdx].type !== 'inline') { continue; }
-      openLink = false;
-      inlineTokens = state.tokens[blkIdx].children;
-
-      for (i = 0; i < inlineTokens.length; i++) {
-        if (inlineTokens[i].type === 'link_open') { openLink = true; }
-        if (inlineTokens[i].type === 'link_close') { openLink = false; }
-        if (inlineTokens[i].type === 'hashtag_open' && openLink) {
-          inlineTokens[i].type = 'text';
-          inlineTokens[i].content = '';
-        }
-        if (inlineTokens[i].type === 'hashtag_text' && openLink) {
-          inlineTokens[i].type = 'text';
-          inlineTokens[i].content = '#' + inlineTokens[i].content;
-          inlineTokens[i].level--;
-        }
-        if (inlineTokens[i].type === 'hashtag_close' && openLink) {
-          inlineTokens[i].type = 'text';
-          inlineTokens[i].content = '';
-        }
-      }
-    }
-  });
 };