+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Head
 
-* Prevent hashtags being rendered inside of link texts
+* Rewrite to prevent hashtags being rendered inside of link texts
 * Add more specs
 
 # 0.1.0 / 2015-01-08